Wave ok, underwater nope
